API reference

Scheme parameters

HEAAN.Params — Type

Params(; log_polynomial_length::Int=16, log_lo_modulus::Int=300)

HEAAN scheme parameters.

log_polynomial_length determines the length of polynomials used for keys and ciphertexts; larger length is more secure, but also slower.

log_lo_modulus determines the maximum computation resource that a ciphertext can have. Larger values are more secure and allow for longer computation sequences without bootstrapping; but, again, are slower and make ciphertexts take more place in memory.

Keys

HEAAN.SecretKey — Type

SecretKey(rng::AbstractRNG, params::Params)

Secret key, used for decryption. Takes a Params object.

HEAAN.EncryptionKey — Type

EncryptionKey(rng::AbstractRNG, secret_key::SecretKey)

A public key used for encryption. Needs a SecretKey object.

HEAAN.MultiplicationKey — Type

MultiplicationKey(rng::AbstractRNG, secret_key::SecretKey)

A public key used for multiplication of two ciphertexts. Needs a SecretKey object.

HEAAN.LeftRotationKey — Type

LeftRotationKey(rng::AbstractRNG, secret_key::SecretKey, shift::Int)

A public key used for left rotation (circshift() with a negative argument) of a ciphertext. Needs a SecretKey object and the absolute value of the shift (a separate key is needed for each value of the shift).

HEAAN.ConjugationKey — Type

LeftRotationKey(rng::AbstractRNG, secret_key::SecretKey)

A public key used for conjugation of a ciphertext. Needs a SecretKey object.

Encryption/decryption

HEAAN.Ciphertext — Type

A ciphertext object created by encrypt.

HEAAN.encrypt — Function

encrypt(
    rng::AbstractRNG, key::EncryptionKey, vals::Array{Complex{Float64}, 1},
    log_precision::Int, log_cap::Int)

Creates a Ciphertext object out of a complex-valued array vals.

log_precision determines the (absolute) precision used for encoding (cannot be larger than log_cap).

log_cap is the available computation resource (cannot be larger than Params.log_lo_modulus; see explanation thereof).

HEAAN.decrypt — Function

decrypt(secret_key::SecretKey, cipher::Ciphertext)

Decrypts a Ciphertext object and returns the resulting complex-valued array.

Arithmetic operations

HEAAN.add — Function

add(cipher1::Ciphertext, cipher2::Ciphertext)

Elementwise addition of encrypted vectors. Both ciphertexts must have the same log_cap and log_precision.

HEAAN.add_const — Function

add_const(cipher::Ciphertext, cnst::Float64)

Adds a floating-point number cnst to each element of an encrypted vector. Ciphertext's log_precision is used to encode cnst.

add_const(cipher::Ciphertext, cnst::Float64)

Adds a complex number cnst to each element of an encrypted vector. Ciphertext's log_precision is used to encode cnst.

HEAAN.sub — Function

sub(cipher1::Ciphertext, cipher2::Ciphertext)

Elementwise subtraction of encrypted vectors. Both ciphertexts must have the same log_cap and log_precision.

HEAAN.mul — Function

mul(key::MultiplicationKey, cipher1::Ciphertext, cipher2::Ciphertext)

Elementwise multiplication of encrypted vectors. Needs a MultiplicationKey. Both ciphertexts must have the same log_cap (but can have different log_precision); log_precision of the result is the sum of log_precision of both ciphertexts.

HEAAN.square — Function

square(mk::MultiplicationKey, cipher::Ciphertext)

Elementwise square of an encrypted vector. Needs a MultiplicationKey object. log_precision of the result is doubled.

HEAAN.negate — Function

negate(cipher::Ciphertext)

Elementwise negation of an encrypted vector.

HEAAN.imul — Function

imul(cipher::Ciphertext)

Multiplies each element of an encrypted vector by the imaginary unit.

HEAAN.mul_by_const — Function

mul_by_const(cipher::Ciphertext, cnst::Float64, log_precision::Int)

Multiplies each element of an encrypted vector by a floating-point number cnst. The provided log_precision is used to encode cnst.

mul_by_const(cipher::Ciphertext, cnst::Complex{Float64}, log_precision::Int)

Multiplies each element of an encrypted vector by a complex number cnst. The provided log_precision is used to encode cnst.

HEAAN.mul_by_const_vec — Function

mul_by_const_vec(cipher::Ciphertext, v::Array{Complex{Float64}, 1}, log_precision::Int)

Multiplies each element of an encrypted vector by the corresponding element of v. log_precision is used to encode elements of v.

HEAAN.div_by_po2 — Function

div_by_po2(cipher::Ciphertext, bits::Int)

Elementwise division of an encrypted vector by 2^bits.

Base.circshift — Function

circshift(rk::LeftRotationKey, cipher::Ciphertext, shift::Integer)

Rotates an encrypted vector (currently only to the left, that is shift must be negative). The LeftRotationKey must correspond to shift.

Base.conj — Function

conj(ck::ConjugationKey, cipher::Ciphertext)

Conjugates an encrypted vector. Needs a ConjugationKey object.

Base.inv — Function

inv(mk::MultiplicationKey, cipher::Ciphertext, log_precision::Int, steps::Int)

Calculates an inverse (1/x) of each element of an encrypted vector, using a steps-power approximation. log_precision determines the step used to rescale the intermediate values. Needs a MultiplicationKey object.

HEAAN.power — Function

power(mk::MultiplicationKey, cipher::Ciphertext, log_precision::Int, pwr::Int)

Raises each element of an encrypted vector to the (non-negative) power pwr. log_precision determines the step used to rescale the intermediate values. Needs a MultiplicationKey object.

HEAAN.power_series — Function

power_series(
    mk::MultiplicationKey, cipher::Ciphertext, log_precision::Int,
    coeffs::Array{Float64, 1}; lazy::Bool=false)

Calculates a power series (a + bx + cx^2 + ...) for each element x of an encrypted vector, with coefficients coeffs. log_precision is used to rescale the intermediate values. Needs a MultiplicationKey object.

HEAAN.sigmoid — Function

sigmoid(mk::MultiplicationKey, cipher::Ciphertext, log_precision::Int, max_pwr::Int)

Calculates the sigmoid function exp(x) / exp(1+x) for each element of an encrypted vector, using a max_pwr-power approximation. log_precision determines the step used to rescale the intermediate values. Needs a MultiplicationKey object.

HEAAN.log_plus_one — Function

log_plus_one(mk::MultiplicationKey, cipher::Ciphertext, log_precision::Int, max_pwr::Int)

Calculates log(1 + x) for each element of an encrypted vector, using a max_pwr-power approximation. log_precision determines the step used to rescale the intermediate values. Needs a MultiplicationKey object.

Base.exp — Function

exp(mk::MultiplicationKey, cipher::Ciphertext, log_precision::Int, max_pwr::Int)

Calculates exp(x) for each element of an encrypted vector, using a max_pwr-power approximation. log_precision determines the step used to rescale the intermediate values. Needs a MultiplicationKey object.

Computation resource management

HEAAN.mod_down_by — Function

mod_down_by(cipher::Ciphertext, dlog_cap::Int)

Decreases log_cap of a Ciphertext object by dlog_cap.

HEAAN.mod_down_to — Function

mod_down_to(cipher::Ciphertext, log_cap::Int)

Decreases log_cap of a Ciphertext object to the provided value.

HEAAN.rescale_by — Function

rescale_by(cipher::Ciphertext, dlog_cap::Int)

Decreases log_cap and log_precision of a Ciphertext object by dlog_cap.

HEAAN.BootstrapKey — Type

BootstrapKey(
    rng::AbstractRNG, secret_key::SecretKey,
    enc_key::EncryptionKey, mul_key::MultiplicationKey, conj_key::ConjugationKey,
    log_slots::Int, log_precision::Int)

A public key used for bootstrapping. Can only be applied to encrypted vector of size 2^log_slots. Needs the SecretKey, an EncryptionKey, a MultiplicationKey and a ConjugationKey.

HEAAN.bootstrap — Function

bootstrap(bk::BootstrapKey, cipher::Ciphertext, log_t::Int=4)

Bootstrap a ciphertext, increasing the difference between its log_cap and log_precision. Needs a BootstrapKey object.